still working on room selection view, starting to try tab approach
